refactor: Moving header and top header to components/general
